Spark: Refactor action for removing orphan files
